refactor(settings): move the SwiftUI color-scheme bridge to the view layer
Lets SettingsManager drop its SwiftUI dependency. (F-ARCH-012)
